+++ Core/Models/Message/StrategyMessage.cs
@@ -7,7 +7,7 @@ public class StrategyMessage
     public StrategyTaskEnum StrategyTask { get; set; } = StrategyTaskEnum.None;
     public bool IsBacktest { get; set; } = false;
     public StrategyEnum Strategy { get; set; } = StrategyEnum.None;
-    public MessageTypeEnum MessageType { get; set; } = MessageTypeEnum.Init;
+    public MessageTypeEnum MessageType { get; set; } = MessageTypeEnum.Start;
     public StrategySettingsModel? Settings { get; set; } = null;
     public List<Quote>? Quotes { get; set; } = null;
 }